split websocket from test (#259)
